+++ src/console/controllers/FlagsController.php
@@ -109,6 +109,9 @@ class FlagsController extends Controller
         $this->stdout(Craft::t('feature-flags', 'Rollout Percentage') . ': ', Console::BOLD);
         $this->stdout(($flag->rolloutPercentage !== null ? $flag->rolloutPercentage . '%' : '—') . PHP_EOL);
 
+        $this->stdout(Craft::t('feature-flags', 'Rollout Strategy') . ': ', Console::BOLD);
+        $this->stdout($flag->rolloutStrategy . PHP_EOL);
+
         $this->stdout(Craft::t('feature-flags', 'Expires At') . ': ', Console::BOLD);
         $this->stdout(($flag->expiresAt ? $flag->expiresAt->format('Y-m-d H:i:s') : '—') . PHP_EOL);